deploy script for Minter.sol
